pyproject.toml
[project]
name = "finsilio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "apscheduler==3.11.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "curl-cffi==0.12.0",
    "fastapi==0.116.1",
    "frozendict==2.4.6",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.177.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.28.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.27",
    "langchain-core==0.3.72",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.6.3",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.3",
    "langgraph-sdk==0.2.0",
    "langsmith==0.4.10",
    "multitasking==0.0.12",
    "numpy==2.3.2",
    "orjson==3.11.1",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "peewee==3.18.2",
    "platformdirs==4.3.8",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-telegram-bot==22.3",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "tenacity==8.5.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "yfinance==0.2.65",
    "zstandard==0.23.0",
]
